Move state type, fix comment
