+++ apps/src/aichat/redux/index.ts
@@ -1,8 +1,8 @@
 export * from './selectors';
 export * from './thunks';
+export type {AichatState} from './state';
 // Only export a subset of actions that are meant to be used by components/application code
 export {
-  AichatState,
   clearChatMessages,
   endSave,
   removeUpdateMessage,
