+++ apps/src/aichat/redux/slice.ts
@@ -7,7 +7,6 @@ import {
   AiCustomizations,
   ChatEvent,
   ChatMessage,
-  FieldVisibilities,
   LevelAichatSettings,
   ModelCardInfo,
   SaveType,
@@ -24,36 +23,7 @@ import {
 } from '../views/modelCustomization/constants';
 import {validateModelId} from '../views/modelCustomization/utils';
 
-const messageListKeys = ['chatEventsPast', 'chatEventsCurrent'] as const;
-type MessageLocation = {
-  index: number;
-  messageListKey: (typeof messageListKeys)[number];
-};
-type ModalType = ModalTypes.WARNING | ModalTypes.TEACHER_ONBOARDING | undefined;
-
-export interface AichatState {
-  // Content from previous chat sessions that we track purely for visibility to the user
-  // and do not send to the model as history.
-  chatEventsPast: ChatEvent[];
-  // Items in the current chat session that we want to provide as history to the model.
-  chatEventsCurrent: ChatEvent[];
-  // The user message currently awaiting response from the model (if any).
-  chatMessagePending?: ChatMessage;
-  // Student events viewed by a teacher user in chat workspace
-  studentChatHistory: ChatEvent[];
-  // Denotes whether we should show the warning or teacher onboarding modal
-  showModalType: ModalType;
-  initialAiCustomizations: AiCustomizations;
-  currentAiCustomizations: AiCustomizations;
-  savedAiCustomizations: AiCustomizations;
-  fieldVisibilities: FieldVisibilities;
-  viewMode: ViewMode;
-  // If a save is currently in progress
-  saveInProgress: boolean;
-  // The type of save action being performed (customization update, publish, model card save, etc).
-  currentSaveType: SaveType | undefined;
-  userHasAichatAccess: boolean;
-}
+import {AichatState} from './state';
 
 const initialState: AichatState = {
   chatEventsPast: [],
@@ -120,7 +90,10 @@ const aichatSlice = createSlice({
       state.chatEventsPast.push(...state.chatEventsCurrent);
       state.chatEventsCurrent = [];
     },
-    setShowModalType: (state, action: PayloadAction<ModalType>) => {
+    setShowModalType: (
+      state,
+      action: PayloadAction<ModalTypes | undefined>
+    ) => {
       state.showModalType = action.payload;
     },
     setViewMode: (state, action: PayloadAction<ViewMode>) => {
@@ -238,10 +211,10 @@ const aichatSlice = createSlice({
   },
 });
 
-const getUpdateMessageLocation = (
-  id: number,
-  state: AichatState
-): MessageLocation | undefined => {
+// List keys of chat events to look through when removing a message.
+const messageListKeys = ['chatEventsPast', 'chatEventsCurrent'] as const;
+
+const getUpdateMessageLocation = (id: number, state: AichatState) => {
   for (const messageListKey of messageListKeys) {
     const messageList = state[messageListKey];
